src/main/java/guru/springframework/controllers/RecipeController.java
package guru.springframework.controllers;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;

import guru.springframework.services.RecipeService;

@Controller
public class RecipeController {
	
	private final RecipeService recipeService;
	
	@Autowired
	public RecipeController(RecipeService recipeService) {
		this.recipeService = recipeService;
	}
	
	@RequestMapping(value="/recipe/show/{id}")
	public String showRecipe(Model model, @PathVariable("id") Long id) {
		model.addAttribute("recipe", recipeService.getRecipeById(id) );
		return "recipe/show";
	}

}
